week1-python-ai-basics/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-chatbot"
version = "0.1.0"
description = "命令行 AI 聊天机器人"
requires-python = ">=3.10"
dependencies = [
    "anthropic",
    "python-dotenv",
    "httpx",
]

[project.scripts]
ai-chatbot = "chatbot.__main__:main"

[tool.setuptools.packages.find]
include = ["chatbot*"]
